Transition to short from int. Added structs.h with the structs that will be used from the functions.

+++ colors.c
@@ -1,7 +1,7 @@
 #include <curses.h>
 #include "colors.h"
 
-void inizializzaColore(int N, int r, int g, int b)
+void inizializzaColore(short N, short r, short g, short b)
 {
     init_color(N, r * CONVERSION_CONSTANT, g * CONVERSION_CONSTANT, b * CONVERSION_CONSTANT);
 }
